requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
requests
